backend/data_enrichment/requirements.txt
# Basic HTTP and utility libraries
requests>=2.28.0
python-dotenv>=0.20.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.1

# Data processing
python-levenshtein>=0.12.2
fuzzywuzzy>=0.18.0
python-dateutil>=2.8.2
pandas>=1.4.2

# Geocoding and location services
geopy>=2.3.0
# Added for multi-provider geocoding
mapbox>=0.18.1  
googlemaps>=4.10.0

# Database
supabase>=0.7.1
neo4j>=5.2.0

# Async task processing
celery>=5.2.7

# LLM APIs (for advanced research)
openai>=0.27.0
anthropic>=0.2.8

# Financial data APIs
yfinance>=0.1.70
fredapi>=0.5.0
alpha_vantage>=2.3.1
polygon-api-client>=1.2.0

# Caching 
diskcache>=5.4.0
cachetools>=5.3.1

# Utilities
tenacity>=8.2.2  # For retry functionality
pytz>=2023.3
aiofiles>=23.1.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0